Add auth validation to request processing
Before this change, requests (e.g. publish) sent by a client after an
unsuccessful authentication were still processed despite the mis-authentication
if they were sent quickly enough before the connection was closed.
This change adds validation checks for successful authentication to all the
processing callbacks so that following requests are ignored in case of an
unsuccessful authentication when connection in the beginning.